utils: cplay: Reset file cursor after MP3 header parse
We need to reset the file cursor to the beginning of the file.

Initially, the program would simply get stuck polling the
compress fd. This was probably because of the fact that the
codec would hang because of the fact that it was expecting
to receive the MP3 data along with its associated MP3 header.

This was not the case for the first (header, data) pair because,
after parsing the first header, the file cursor would point
at the beginning of the data region.

By resetting the file cursor to the beginning of the file,
the codec will receive all the (header, data) pairs it
actually expects.

Suggested-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>